tok ut en haug med funksjoner fra grunnkart så det blir litt lettere å finne ting, så sporet litt av igjen. ops.
