+++ src/Grunnkart/GrunnkartFunksjoner/redirectTo.js
@@ -1,0 +1,4 @@
+export default function redirectTo(path, place) {
+  const newUrl = "/" + path;
+  place.props.history.replace(newUrl);
+}
